8896: combining helpers [skip ci]
