+++ web-client/src/presenter/computeds/TrialSession/addTrialSessionInformationHelper.test.js
@@ -2,14 +2,14 @@ import {
   TRIAL_SESSION_PROCEEDING_TYPES,
   TRIAL_SESSION_SCOPE_TYPES,
 } from '../../../../../shared/src/business/entities/EntityConstants';
+import { addTrialSessionInformationHelper as addTrialSessionInformationHelperComputed } from './addTrialSessionInformationHelper';
 import { applicationContextForClient as applicationContext } from '../../../../../shared/src/business/test/createTestApplicationContext';
 import { runCompute } from 'cerebral/test';
-import { trialSessionInformationHelper as trialSessionInformationHelperComputed } from './trialSessionInformationHelper';
 import { withAppContextDecorator } from '../../../withAppContext';
 
-describe('trialSessionInformationHelper', () => {
-  const trialSessionInformationHelper = withAppContextDecorator(
-    trialSessionInformationHelperComputed,
+describe('addTrialSessionInformationHelper', () => {
+  const addTrialSessionInformationHelper = withAppContextDecorator(
+    addTrialSessionInformationHelperComputed,
     {
       ...applicationContext,
     },
@@ -17,7 +17,7 @@ describe('trialSessionInformationHelper', () => {
 
   describe('displayRemoteProceedingForm', () => {
     it(`should be true when form.proceedingType is ${TRIAL_SESSION_PROCEEDING_TYPES.remote}`, () => {
-      const result = runCompute(trialSessionInformationHelper, {
+      const result = runCompute(addTrialSessionInformationHelper, {
         state: {
           form: { proceedingType: TRIAL_SESSION_PROCEEDING_TYPES.remote },
         },
@@ -27,7 +27,7 @@ describe('trialSessionInformationHelper', () => {
     });
 
     it(`should be true when the session scope is ${TRIAL_SESSION_SCOPE_TYPES.standaloneRemote}`, () => {
-      const result = runCompute(trialSessionInformationHelper, {
+      const result = runCompute(addTrialSessionInformationHelper, {
         state: {
           form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.standaloneRemote },
         },
@@ -37,7 +37,7 @@ describe('trialSessionInformationHelper', () => {
     });
 
     it(`should be false when the session scope is NOT ${TRIAL_SESSION_SCOPE_TYPES.standaloneRemote} and form.proceedingType is NOT ${TRIAL_SESSION_PROCEEDING_TYPES.remote}`, () => {
-      const result = runCompute(trialSessionInformationHelper, {
+      const result = runCompute(addTrialSessionInformationHelper, {
         state: {
           form: { proceedingType: 'def', sessionScope: 'abc' },
         },
@@ -49,7 +49,7 @@ describe('trialSessionInformationHelper', () => {
 
   describe('isStandaloneSession', () => {
     it(`should be true when form.sessionScope is ${TRIAL_SESSION_SCOPE_TYPES.standaloneRemote}`, () => {
-      const result = runCompute(trialSessionInformationHelper, {
+      const result = runCompute(addTrialSessionInformationHelper, {
         state: {
           form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.standaloneRemote },
         },
@@ -59,7 +59,7 @@ describe('trialSessionInformationHelper', () => {
     });
 
     it(`should be false when the session scope is ${TRIAL_SESSION_SCOPE_TYPES.locationBased}`, () => {
-      const result = runCompute(trialSessionInformationHelper, {
+      const result = runCompute(addTrialSessionInformationHelper, {
         state: {
           form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.locationBased },
         },
@@ -71,7 +71,7 @@ describe('trialSessionInformationHelper', () => {
 
   describe('title', () => {
     it(`should be 'Remote Proceeding Information' when form.sessionScope is ${TRIAL_SESSION_SCOPE_TYPES.standaloneRemote}`, () => {
-      const result = runCompute(trialSessionInformationHelper, {
+      const result = runCompute(addTrialSessionInformationHelper, {
         state: {
           form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.standaloneRemote },
         },
@@ -81,13 +81,63 @@ describe('trialSessionInformationHelper', () => {
     });
 
     it(`should be 'Location Information' when form.sessionScope is ${TRIAL_SESSION_SCOPE_TYPES.locationBased}`, () => {
-      const result = runCompute(trialSessionInformationHelper, {
+      const result = runCompute(addTrialSessionInformationHelper, {
         state: {
           form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.locationBased },
         },
       });
 
       expect(result.title).toEqual('Location Information');
+    });
+  });
+
+  describe('isStandaloneSession', () => {
+    it(`should be true when form.sessionScope is ${TRIAL_SESSION_SCOPE_TYPES.standaloneRemote}`, () => {
+      const result = runCompute(addTrialSessionInformationHelper, {
+        state: {
+          form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.standaloneRemote },
+        },
+      });
+
+      expect(result.isStandaloneSession).toEqual(true);
+    });
+
+    it(`should be false when form.sessionScope is ${TRIAL_SESSION_SCOPE_TYPES.locationBased}`, () => {
+      const result = runCompute(addTrialSessionInformationHelper, {
+        state: {
+          form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.locationBased },
+        },
+      });
+
+      expect(result.isStandaloneSession).toEqual(false);
+    });
+  });
+
+  describe('sessionTypes', () => {
+    it(`should NOT include 'Special' or 'Motion/Hearing' when form.sessionScope is ${TRIAL_SESSION_SCOPE_TYPES.standaloneRemote}`, () => {
+      const result = runCompute(addTrialSessionInformationHelper, {
+        state: {
+          form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.standaloneRemote },
+        },
+      });
+
+      expect(result.sessionTypes).toEqual(['Regular', 'Small', 'Hybrid']);
+    });
+
+    it(`should include 'Special' and 'Motion/Hearing' when form.sessionScope is ${TRIAL_SESSION_SCOPE_TYPES.locationBased}`, () => {
+      const result = runCompute(addTrialSessionInformationHelper, {
+        state: {
+          form: { sessionScope: TRIAL_SESSION_SCOPE_TYPES.locationBased },
+        },
+      });
+
+      expect(result.sessionTypes).toEqual([
+        'Regular',
+        'Small',
+        'Hybrid',
+        'Special',
+        'Motion/Hearing',
+      ]);
     });
   });
 });
